fix(test): fix for parse test
exclude assertion of log output since it is a bad practice
